add test_common
